tests

+++ src/geometry/polygon_properties.hpp
@@ -19,13 +19,11 @@ namespace kac_core::geometry {
 
 	inline bool isConvex(const T::Polygon& P) {
 		/*
-		Tests whether or not a given array of vertices forms a convex polygon.
-		This is achieved using the resultant sign of the cross product for each
-		vertex:
+		Tests whether or not a given array of vertices forms a convex polygon. This is achieved
+		using the resultant sign of the cross product for each vertex:
 			[(x_i - x_i-1), (y_i - y_i-1)] × [(x_i+1 - x_i), (y_i+1 - y_i)]
-		See => http://paulbourke.net/geometry/polygonmesh/ 'Determining whether
-		or not a polygon (2D) has its vertices ordered clockwise or
-		counter-clockwise'.
+		See => http://paulbourke.net/geometry/polygonmesh/ 'Determining whether or not a polygon
+		(2D) has its vertices ordered clockwise or counter-clockwise'.
 		*/
 
 		// cross product - z component only, see np.cross =>
@@ -34,7 +32,7 @@ namespace kac_core::geometry {
 			return (p.x - p_minus.x) * (p_plus.y - p.y) - (p_plus.x - p.x) * (p.y - p_minus.y);
 		};
 		// determine the direction of the initial point using the cross product
-		const unsigned long N = static_cast<unsigned long>(P.size());
+		const unsigned long N = P.size();
 		bool clockwise = crossProductZ(P[0], P[1], P[N - 1]) < 0;
 		// loop over remaining points
 		for (unsigned long n = 1; n < N; n++) {
@@ -56,7 +54,7 @@ namespace kac_core::geometry {
 			(P[1].x - P[0].x) * (P[2].y - P[1].y) - (P[1].y - P[0].y) * (P[2].x - P[1].x) > 0 ? -1
 																							  : 1;
 		// go through each of the vertices, plus the next vertex in the list
-		const unsigned long N = static_cast<unsigned long>(P.size());
+		const unsigned long N = P.size();
 		for (unsigned long n = 0; n < N; n++) {
 			T::Point a = P[n];
 			T::Point b = P[(n + 1) % N];
@@ -72,7 +70,7 @@ namespace kac_core::geometry {
 		Determine if a polygon is simple by checking for intersections.
 		*/
 
-		const unsigned long N = static_cast<unsigned long>(P.size());
+		const unsigned long N = P.size();
 		for (unsigned long i = 0; i < N - 2; i++) {
 			for (unsigned long j = i + 1; j < N; j++) {
 				std::string intersection_type =
@@ -89,11 +87,11 @@ namespace kac_core::geometry {
 
 	inline std::pair<double, std::pair<long, long>> largestVector(const T::Polygon& P) {
 		/*
-		This function tests each pair of vertices in a given polygon to find the
-		largest vector, and returns the length of the vector and its indices.
+		This function tests each pair of vertices in a given polygon to find the largest vector, and
+		returns the length of the vector and its indices.
 		*/
 
-		const unsigned long N = static_cast<unsigned long>(P.size());
+		const unsigned long N = P.size();
 		double vec_max = 0.;
 		long index_i = 0;
 		long index_j = 0;
@@ -116,7 +114,7 @@ namespace kac_core::geometry {
 		https://math.blogoverflow.com/2014/06/04/greens-theorem-and-area-of-polygons/
 		*/
 
-		const unsigned long N = static_cast<unsigned long>(P.size());
+		const unsigned long N = P.size();
 		double out = 0.;
 		for (unsigned long n = 0; n < N; n++) {
 			out += (P[(n + 1) % N].x + P[n].x) * (P[(n + 1) % N].y - P[n].y);
@@ -126,10 +124,9 @@ namespace kac_core::geometry {
 
 	inline T::Point polygonCentroid(const T::Polygon& P, const double& area) {
 		/*
-		This algorithm is used to calculate the geometric centroid of a 2D
-		polygon. See http://paulbourke.net/geometry/polygonmesh/ 'Calculating
-		the area and centroid of a polygon'.
-
+		This algorithm is used to calculate the geometric centroid of a 2D polygon.
+		See http://paulbourke.net/geometry/polygonmesh/ 'Calculating the area and centroid of a
+		polygon'.
 		output:
 			for N == 3 ->
 			(x, y) = (
@@ -143,7 +140,7 @@ namespace kac_core::geometry {
 			)
 		 */
 
-		const unsigned long N = static_cast<unsigned long>(P.size());
+		const unsigned long N = P.size();
 		double out_x = 0.;
 		double out_y = 0.;
 		if (N == 3) {
